ai/requirements.txt
fastapi
uvicorn
pymongo
openai
scikit-learn
sentence-transformers
python-dotenv